refactor(components): split RouteCard into modular subcomponents (P2-2)
RouteCard.tsx had grown to 1,256 lines holding the card shell plus ten
sub-components and the pure helpers behind them. Move each concern into its
own module under shared/RouteCard/ and keep an index barrel so the existing
"@/components/shared/RouteCard" import path and its exported helpers
(shouldAppendExitNumber, getConfidenceLabelKey, getRouteAlertsCount,
dedupeA11yCategories) stay identical.
