+++ src/components/shared/RouteCard/index.ts
@@ -1,0 +1,7 @@
+export { RouteCard } from "./RouteCard";
+export {
+  dedupeA11yCategories,
+  getConfidenceLabelKey,
+  getRouteAlertsCount,
+  shouldAppendExitNumber,
+} from "./utils";
